+++ lib/clickhouse-activerecord/tasks.rb
@@ -36,14 +36,14 @@ module ClickhouseActiverecord
 
       File.open(args.first, 'w:utf-8') do |file|
         tables.each do |table|
-          file.puts connection.execute("SHOW CREATE TABLE #{table}")['data'].try(:first).try(:first).gsub("#{@configuration['database']}.", '')
-          file.puts '--- SEPARATOR'
+          next if table.match(/\.inner/)
+          file.puts connection.execute("SHOW CREATE TABLE #{table}")['data'].try(:first).try(:first).gsub("#{@configuration['database']}.", '') + ";\n\n"
         end
       end
     end
 
     def structure_load(*args)
-      File.read(args.first).split("\n--- SEPARATOR\n").each { |sql| connection.execute(sql) }
+      File.read(args.first).split(";\n\n").each { |sql| connection.execute(sql) }
     end
 
     def migrate